Move LogicalPlanBuilder into builder.rs
